Moved method GetIsoYearAndWeek to Utilityfolder
